Add missing packages
The electron packages were not required dependencies and thus a clean
installation wouldn't work. The `start` script was also missing that was
a part of the instructions.

The addition of `nodeIntegration: true` seems to make the code
work. This probably should've been included in the code in the first
place as axios was used in the browser context.